requirements.txt
Flask==2.0.1
Flask-CORS==3.0.10
Werkzeug==2.0.1
Pillow==8.2.0
google-generativeai==0.3.1
ebaysdk==2.2.0
